bugfix for attackorder described in issue#28 https://github.com/optc-db/optc-db.github.io/issues/28
given the case the attackOrder is not preset, it is determined via sorting by highest damage of each unit.
however, this happens BEFORE the cptEffects are applied! so if a unit is not affected by the cptEffects but has a higher BaseDamage than a unit which is affected by cptEffects it will most likely have assigned a higher attackOrder position than it actually should have to get the maximum overallDamage.

my quickfix is simply to redo the sorting after the cptEffects are applied.